requirements.txt
Django==5.2.6
 djangorestframework==3.16.1
 django-filter==25.1
 drf-nested-routers==0.95.0
 drf-spectacular==0.27.0
 djangorestframework_simplejwt==5.5.1
 dj-rest-auth==7.0.1
 django-allauth==65.3.0
 django-cors-headers==4.8.0
 psycopg2-binary==2.9.9
 django-environ==0.11.2
 python-dotenv==1.0.1
 coverage==7.6.1
 requests==2.32.3
flake8==7.3.0